+++ Java/MADJ/src/main/java/com/madj/ProductRepository.java
@@ -1,5 +1,5 @@
 package com.madj;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface ProductRepository extends JpaRepository<Product, Long>  {
+interface ProductRepository extends JpaRepository<Product, Integer>  {
 }
